Add button animations
When the button is clicked we add the `playing` class to it, which
provides an animation to highlight and enlarge the button. In order to
allow the animation to complete before removing the class, we remove it
in a `transitionend` callback.

The styles provided for this use `transition: all`, so we need to tell
it to only run the callback for one transition to avoid executing it
about 7 times.